+++ src/main/java/com/community/tools/service/StateMachineService.java
@@ -126,92 +126,4 @@ public class StateMachineService {
       return false;
     }
   }
-
-  /**
-   * Method to start the event by state.
-   *
-   * @param messageFromUser message from user
-   * @param userId user id
-   * @throws Exception Exception
-   */
-  public void doAction(String messageFromUser, String userId) throws Exception {
-
-    StateMachine<State, Event> machine = restoreMachine(userId);
-    String userForQuestion = machine.getExtendedState().getVariables().get("id").toString();
-
-    String message = Messages.DEFAULT_MESSAGE;
-    Event event = null;
-    Payload payload = null;
-
-    switch (machine.getState().getId()) {
-      case NEW_USER:
-        if (messageFromUser.equalsIgnoreCase("ready")) {
-          payload = new SinglePayload(userId);
-          event = Event.QUESTION_FIRST;
-        } else {
-          message = Messages.NOT_THAT_MESSAGE;
-        }
-        break;
-      case FIRST_QUESTION:
-        payload = new QuestionPayload(userId, messageFromUser, userForQuestion);
-        event = Event.QUESTION_SECOND;
-        break;
-      case SECOND_QUESTION:
-        payload = new QuestionPayload(userId, messageFromUser, userForQuestion);
-        event = Event.QUESTION_THIRD;
-        break;
-      case THIRD_QUESTION:
-        payload = new QuestionPayload(userId, messageFromUser, userForQuestion);
-        event = Event.CONSENT_TO_INFORMATION;
-        break;
-      case AGREED_LICENSE:
-        String gitNick = messageFromUser;
-        payload = new VerificationPayload(userId, gitNick);
-        event = Event.LOGIN_CONFIRMATION;
-        break;
-      case CHECK_LOGIN:
-        if (messageFromUser.equalsIgnoreCase("yes")) {
-          event = Event.ADD_GIT_NAME_AND_FIRST_TASK;
-        } else if (messageFromUser.equalsIgnoreCase("no")) {
-          event = Event.DID_NOT_PASS_VERIFICATION_GIT_LOGIN;
-        } else {
-          message = Messages.NOT_THAT_MESSAGE;
-        }
-        payload = (VerificationPayload) machine.getExtendedState().getVariables()
-            .get("dataPayload");
-        break;
-      default:
-        event = null;
-        payload = null;
-    }
-    if (event == null) {
-      messageService.sendPrivateMessage(messageService.getUserById(userId), message);
-    } else {
-      machine.getExtendedState().getVariables().put("dataPayload", payload);
-      machine.sendEvent(event);
-      persistMachine(machine, payload.getId());
-    }
-  }
-
-  /**
-   * Reset User with userId.
-   *
-   * @param userId platform user id
-   * @throws Exception Exception
-   */
-  public void resetUser(String userId) throws Exception {
-
-    User stateEntity = new User();
-    stateEntity.setUserID(userId);
-    stateMachineRepository.save(stateEntity);
-    persistMachineForNewUser(userId);
-
-    String user = messageService.getUserById(userId);
-    messageService.sendPrivateMessage(user,
-        Messages.WELCOME);
-    messageService
-        .sendBlocksMessage(user, messageService
-            .createBlockMessage(MessagesToSlack.MESSAGE_ABOUT_RULES,
-                MessagesToDiscord.MESSAGE_ABOUT_RULES));
-  }
 }
